requirements.txt
streamlit==1.52.1
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.0.0
pandas==2.3.3
h5py==3.15.1
protobuf==6.33.2
keras==3.12.0
